Move messages into protocol, documents messages
