+++ src/lib.rs
@@ -3,8 +3,8 @@
 pub mod blockchains;
 pub mod crypto;
 pub mod instructions;
-pub mod messages;
 pub mod negotiation;
+pub mod protocol;
 pub mod roles;
 pub mod session;
 pub mod version;
